requirements.txt
allure-pytest==2.10.0
allure-python-commons==2.10.0
Appium-Python-Client==2.2.0
applescript==2021.2.9
appscript==1.1.2
async-generator==1.10
attrs==20.3.0
beautifulsoup4==4.10.0
cachetools==5.0.0
casefy==0.1.2
certifi==2022.12.7
cffi==1.15.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
colorlog==5.0.1
ConfigArgParse==1.5.3
configparser==5.0.2
crayons==0.4.0
cryptography==36.0.2
et-xmlfile==1.1.0
Flask==1.1.2
Flask-BasicAuth==0.2.0
future==0.18.2
gevent==21.1.2
geventhttpclient==1.4.4
google==3.0.0
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
greenlet==1.1.0
gspread==5.2.0
h11==0.13.0
httplib2==0.20.4
humanfriendly==10.0
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
junitparser==2.8.0
locust==1.5.3
logger==1.4
MarkupSafe==2.0.1
msgpack==1.0.2
mypy-extensions==0.4.3
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.10
outcome==1.1.0
packaging==20.9
pluggy==0.13.1
protobuf==3.19.4
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive==1.3.1
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyserde==0.9.2
PySocks==1.7.1
pytest==7.1.2
pytest-html==3.1.1
pytest-metadata==1.11.0
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.8
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
trcli==1.3.0
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==4.3.0
typing-inspect==0.8.0
uritemplate==4.1.1
urllib3==1.26.4
webdriver-manager==3.4.2
Werkzeug==2.0.1
wsproto==1.1.0
zope.event==4.5.0
zope.interface==5.4.0
